requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
langchain
torch
PyMuPDF               # for PDF reading
